python-service/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
httpx
pytest 
python-dotenv